+++ packages/itwinui-react/src/core/Table/columns/actionColumn.tsx
@@ -5,7 +5,7 @@
 import * as React from 'react';
 import type { HeaderProps } from '../../../react-table/react-table.js';
 import { Checkbox } from '../../Checkbox/Checkbox.js';
-import { SvgColumnManager } from '../../utils/index.js';
+import { SvgColumnManager } from '../../../utils/index.js';
 import { DropdownMenu } from '../../DropdownMenu/DropdownMenu.js';
 import type { DropdownMenuProps } from '../../DropdownMenu/DropdownMenu.js';
 import { IconButton } from '../../Buttons/IconButton.js';
@@ -86,7 +86,7 @@ export const ActionColumn = <T extends Record<string, unknown>>({
               // and table is scrolled to the very left which means our visibility dropdown menu is not visible.
               // So for better UX we need to scroll to that dropdown menu.
               queueMicrotask(() => {
-                buttonRef.current?.scrollIntoView();
+                buttonRef.current?.scrollIntoView({ block: 'nearest' });
               });
             };
             return (