+++ packages/itwinui-react/src/core/VisuallyHidden/VisuallyHidden.tsx
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 import * as React from 'react';
 import cx from 'classnames';
-import type { PolymorphicForwardRefComponent } from '../utils/props.js';
-import { Box, ShadowRoot, useIsClient } from '../utils/index.js';
+import type { PolymorphicForwardRefComponent } from '../../utils/props.js';
+import { Box, ShadowRoot, useHydration } from '../../utils/index.js';
 
 type VisuallyHiddenOwnProps = {
   /**
@@ -36,7 +36,7 @@ export const VisuallyHidden = React.forwardRef((props, ref) => {
     ...rest
   } = props;
 
-  const isMounted = useIsClient();
+  const isHydrated = useHydration() === 'hydrated';
 
   // ShadowRoot is not supported on all elements, so we only use it for few common ones.
   const children = !['div', 'span', 'p'].includes(asProp) ? (
@@ -47,9 +47,9 @@ export const VisuallyHidden = React.forwardRef((props, ref) => {
         <slot />
       </ShadowRoot>
 
-      {/* Render childrenProp only after ShadowRoot attaches the shadow DOM (i.e. only after the first frame) */}
+      {/* Render childrenProp only after ShadowRoot attaches the shadow DOM (i.e. after hydration) */}
       {/* See: https://github.com/iTwin/iTwinUI/issues/1930 */}
-      {isMounted && childrenProp}
+      {isHydrated && childrenProp}
     </>
   );